move all dummy factory and classes in tests.mocking.dummy
