--- tests/mocking/dummy/reverse_proxy.py
@@ -1,32 +1,0 @@
-from remoteappmanager.services.reverse_proxy import ReverseProxy
-from tests.utils import mock_coro_factory
-
-
-def create_reverse_proxy(params=None,
-                         register_result='',
-                         unregister_result=''):
-    """ Return a ReverseProxy with the given parmaters
-
-    Parameters
-    ----------
-    params: dict
-       keyword arguments for initialising ReverseProxy
-
-    register_result : str
-       result of ReverseProxy.register
-
-    unregister_result : str
-       result of ReverseProxy.unregister
-
-    Returns
-    -------
-    ReverseProxy : remoteappmanager.services.reverse_proxy.ReverseProxy
-    """
-    if params is None:
-        params = {}
-
-    revproxy = ReverseProxy(**params)
-    revproxy.register = mock_coro_factory(register_result)
-    revproxy.unregister = mock_coro_factory(unregister_result)
-
-    return revproxy
